+++ lib/archiver.js
@@ -25,9 +25,10 @@ function Archiver(root, store) {
 			var stats = dat.trackStats()
 
 			network.on('connection', () => {
-				var data = store.get_temp(url) || {}
+				var data = store.get_temp(url)
 				data.peers = stats.peers.complete
-				store.set_temp(url, data)
+				console.log(dirname + ': ' + stats.peers.complete)
+				store.set_temp(dirname, data)
 			})
 		})
 	}
