+++ lib/router.js
@@ -37,7 +37,7 @@ function r(store, archiver) {
 	})
 
 	router.get('/stats/:feed', async (ctx) => {
-		var stats = store.get_temp('dat://' + ctx.params.feed)
+		var stats = store.get_temp(ctx.params.feed)
 
 		ctx.response.set('Content-Type', 'application/json')
 		ctx.body = {stats: stats}
