+++ lib/store.js
@@ -17,7 +17,7 @@ function Store(path) {
 	}
 
 	this.get_temp = function(key) {
-		return this.temp[key]
+		return this.temp[key] || {}
 	}
 
 	this.set_temp = function(key, value) {
